Added lists
results are added to list when capture is pressed
Idea was to find average when result is fluctuating. not sure how much difference there will be between frames. Got it working but might not be much benefit. Going to work on implementing region of interest on tap next on a different branch.